feat: add autodocs
